plot function

[...]

# S3 method for ineqx
plot(
  ineqx.out,
  type,
  yint = 1,
  xlab = NULL,
  ylab = NULL,
  llab = NULL,
  titl = NULL,
  xlim = NULL,
  hline = F,
  se = F,
  lowess = F,
  legend = T,
  sav = F
)

Arguments

ineqx.out

ineqx.out object from ineqx()

type

Character string. Plot type. Choose from: dMuP, dMuT, dSigmaP, dSigmaT, dWP, dWT, dBP, dBT, dPP, dPT, dT, dPA

yint

Either 1 or 2. Choose reported effect value. 1: Effect in units of ystat. 2: Effect as change in percent since 'cf' (reference time)

xlab

Character string. x-axis label.

ylab

Vector of length 2 with character strings. y-axis label. c("y1lab", "y2lab")

llab

Vector of length equal to the number of levels of groupvar. Group labels. c("grouplbl1", "grouplbl2", "grouplbl3", ...)

titl

Character string of length 2. Plot titles. c("titl1", "titl2")

xlim

Sequence of numbers.breaks at x-axis. seq(2000,2020, 5)

hline

Number or FALSE. Horizontal line at specified value.

se

Logical. Display standard errors?

lowess

Logical. Display lowess smoother?

legend

Logical. Display legend?

sav

Logical. Save as .png in working directory?

Value

Returns a ggplot2 object

Details

Interpretation of the y-axis

The interpretation of yint=1 (default) is in absolute points of CV2{W/B/T}.

The interpretation of yint=2

For plot(ineqx.out, type="dBP") or plot(ineqx.out, type="dBT"), it is the change in the proportion of dB/CV2B, dP/CV2B relative to the reference value given in ref. The interpretation of dB is analogous.

For plot(ineqx.out, type="dT"), it is the change in the proportion of dW/CV2T, dB/CV2T, dP/CV2T, dT/CV2T relative to the reference value given in ref.

In the example, the CV2 is used as inequality statistic. But the interpretation for the variance as inequality statistic is analogous.

Author

Benjamin Rosche <benjamin.rosche@gmail.com>

Examples

data(incdat)
i1 <- ineqx(incdat, by, inc, groupvar = SES, timevar = year)
#> Error in ineqx(incdat, by, inc, groupvar = SES, timevar = year): unused arguments (groupvar = SES, timevar = year)
plot(i1, type="dMuSigma2")
#> Error in plot(i1, type = "dMuSigma2"): object 'i1' not found