Create a bootstrap configuration

Groups all bootstrap settings into a single object that can be passed to ineqx via the boot argument.

Usage

boot_config(
  data,
  formula_mu,
  formula_sigma,
  treat,
  group = NULL,
  time = NULL,
  post = NULL,
  B = 200L,
  family = NULL,
  parallel = FALSE,
  ncores = NULL,
  seed = NULL,
  verbose = TRUE,
  gamlss_args = list()
)

Arguments

data

Data.frame, the original individual-level data used to fit the GAMLSS model

formula_mu

Formula for the mean (mu) equation

formula_sigma

Formula for the log-SD (sigma) equation (one-sided, e.g., ~treat*group)

treat

Character, name of the treatment variable (coded 0/1)

group

Character, name of the grouping variable. Required when using boot_config via ineqx.

time

Character, name of the time variable. NULL for cross-sectional.

post

Character, name of the pre/post indicator for DiD designs. NULL for simple difference estimator.

B

Integer, number of bootstrap replicates. Default 200.

family

A gamlss.family object (e.g., gamlss.dist::NO()). If NULL, uses the normal distribution.

parallel

Logical, use parallel computation. Default FALSE.

ncores

Integer, number of cores for parallel. Default: all but one.

seed

Integer, random seed for reproducibility. Default NULL.

verbose

Logical, print progress messages. Default TRUE.

gamlss_args

Named list of additional arguments passed to gamlss::gamlss().

Value

An object of class "ineqx_boot_config"

Examples

if (FALSE) { # \dontrun{
bc <- boot_config(
  data = mydata,
  formula_mu = y ~ treat * group,
  formula_sigma = ~ treat * group,
  treat = "treat",
  B = 500,
  seed = 42
)
result <- ineqx(params, se_method = "bootstrap", boot = bc)
} # }